List the configured devices from the EI studio. Migrated to Google accompanist. Improves the response handling of EI Api calls.

+++ app/src/main/java/no/nordicsemi/android/ei/model/Sensor.kt
@@ -3,5 +3,5 @@ package no.nordicsemi.android.ei.model
 data class Sensor(
     val name: String,
     val maxSampleLengths: Int,
-    val frequencies: List<Int>
+    val frequencies: List<Double>
 )
